Add SyncState#update
This replaces #mergeFrom as the public api and takes us from an imperative model to a declarative one. Allowing doOp to force ops to be included or excluded in realtime, and requiring it to also skip ops that were not allowed, prevented us from applying ops that should be applied according to later rules. So, for example, you couldn't remove a writer and then add it back, since any ops after it was removed would be rejected, in spite of the later AddWriter op.