Switch to undoing desired and actual heads
Previously, we looked for the common point between the new and old versions of a head, and then undid the state back to that point. That scheme had many more places where we looked at the device id of an op, making it harder to allow for ops to be merged in from non-writer devices.