Replace our tree of shared nodes with a map of shared nodes
This enables us to represent a shared node even when we don't know where it lives in the larger tree, which happens if we have a shared node that contains ops that come before the op that connects it to the parent that we know about.

For some reason, this made the monkey test 2-3x slower. I think/hope that this is because of more uses of ObjectValue but I didn't profile.