ip. Attempt to turn everything into a connected tree
The problem is that things are not always connected:
1. It's possible now to add a shared node into another shared node after there are earlier ops, so the share-tree isn't always connected.
2. In the future, when we allow moves between shared nodes, the same thing can happen inside a shared node.

The plan was to deal with problem 1 by keeping a map with all of the shared nodes, in addition to having the connected tree. Problem 2 suggests that we'd need to put everything in the map, or continue with the two level map that's currently in master. Or we get fancier and have a forest, ie we only store the disconnected roots of the tree, which would make for more complicated code but a smaller tree.