Start tracking heads per node
So far this is only on UpNode. The goal is to be able to use this to generate the heads on a move op.

Since we like to compare heads using ControlledOpSet.headsEqual (which checks the identity of the linked list) we needed to thread the heads (not just the stream ids and op) through from COS to doOp so that it can use the existing list.