Fix up DynamicPermGroup.equals
We need to ignore the `heads` field because it contains a value that doesn't have true equality (in the Prelude sense). We were getting away with this before because the objects that we compared differed in the `edges` field and `edges` comes before `heads` lexicographically. When I tried to rename `edges` to `writers` things broke. Ugh.